fix import, rename test file
